Refactor the executor (#111)
* Refactored Virtual machine to remove data structure specific functions to their own packages

Linting

* Separated out executor values for each data structure

* Removed priority to fix ordering for expanding (#114)

* Produce Debain files for packaging (#113)

* ubuntu packaging first version

* Added bin files

* Add contributing and code of conduct to repo (#112)

* Added code of conduct

* Added contributing guidelines

* Updated README.md

* Moved DataStructureShapes.kt into datastructures package and moved MakeConstructorNode.kt contents into DataStructureExecutor.kt file

* Refactored if statements

* Removed duplication from for and while loops

* Restructured VirtualMachine.kt file
