+++ src/main/kotlin/com/manimdsl/runtime/ExecValue.kt
@@ -3,13 +3,10 @@ package com.manimdsl.runtime
 import com.manimdsl.frontend.ErrorType
 import com.manimdsl.linearrepresentation.EmptyMObject
 import com.manimdsl.linearrepresentation.MObject
-import com.manimdsl.stylesheet.AnimationProperties
-import com.manimdsl.stylesheet.StyleProperties
-import java.util.*
 import kotlin.math.roundToInt
 
 // Wrapper classes for values of variables while executing code
-sealed class ExecValue {
+abstract class ExecValue {
     abstract var manimObject: MObject
     abstract val value: Any
     abstract val name: String
@@ -120,128 +117,6 @@ data class BoolValue(override val value: Boolean, override var manimObject: MObj
 
     override fun clone(): ExecValue {
         return BoolValue(value, manimObject)
-    }
-}
-
-data class StackValue(override var manimObject: MObject, val stack: Stack<ExecValue>, var style: StyleProperties = StyleProperties(), var animatedStyle: AnimationProperties? = null) : ExecValue() {
-    override val value: Stack<ExecValue> = stack
-    override val name: String = "Stack"
-    override fun clone(): ExecValue {
-        return StackValue(manimObject, stack, style, animatedStyle)
-    }
-
-    override fun toString(): String {
-        return stack.joinToString(" -> ", "[", "]")
-    }
-}
-
-data class ArrayValue(override var manimObject: MObject, val array: Array<ExecValue>, var style: StyleProperties = StyleProperties(), var animatedStyle: AnimationProperties? = null) : ExecValue() {
-    override val value: Array<ExecValue> = array
-    override val name: String = "Array"
-    override fun clone(): ExecValue {
-        return ArrayValue(manimObject, array, style, animatedStyle)
-    }
-
-    override fun toString(): String {
-        return array.joinToString(", ", "[", "]")
-    }
-}
-
-sealed class ITreeNodeValue : ExecValue()
-
-object NullValue : ITreeNodeValue() {
-
-    override val name: String = "Null"
-    override var manimObject: MObject = EmptyMObject
-    override val value: Int = 0
-
-    override fun clone(): ExecValue {
-        return this
-    }
-
-    override fun toString(): String {
-        return "()"
-    }
-}
-
-fun nodeCount(node: ITreeNodeValue): Int {
-    return if (node is NullValue) {
-        0
-    } else {
-        1 + nodeCount((node as BinaryTreeNodeValue).left) + nodeCount(node.right)
-    }
-}
-
-// value is element in node
-data class BinaryTreeNodeValue(
-    var left: ITreeNodeValue = NullValue,
-    var right: ITreeNodeValue = NullValue,
-    override val value: PrimitiveValue,
-    override var manimObject: MObject = EmptyMObject,
-    var binaryTreeValue: BinaryTreeValue? = null,
-    var pathFromRoot: String = "",
-    var depth: Int
-) : ITreeNodeValue() {
-    override fun clone(): ExecValue {
-        return BinaryTreeNodeValue(left, right, value, manimObject, depth = depth)
-    }
-
-    override val name: String = "Tree"
-    fun attachTree(tree: BinaryTreeValue, prefix: String = "${tree.manimObject.shape.ident}.root") {
-        binaryTreeValue = tree
-        pathFromRoot = prefix
-        if (left is BinaryTreeNodeValue) {
-            (left as BinaryTreeNodeValue).attachTree(tree, "$prefix.left")
-        }
-        if (right is BinaryTreeNodeValue) {
-            (right as BinaryTreeNodeValue).attachTree(tree, "$prefix.right")
-        }
-    }
-
-    override fun toString(): String {
-        val stringBuilder = StringBuilder(value.toString())
-        if (left is NullValue) {
-            stringBuilder.append(" $left")
-        } else {
-            stringBuilder.append(" (${left.value}...)")
-        }
-        if (right is NullValue) {
-            stringBuilder.append(" $right")
-        } else {
-            stringBuilder.append(" (${right.value}...)")
-        }
-
-        return stringBuilder.toString()
-    }
-}
-
-// value is root
-data class BinaryTreeValue(override var manimObject: MObject, override var value: BinaryTreeNodeValue, var style: StyleProperties = StyleProperties(), var animatedStyle: AnimationProperties? = null) : ExecValue() {
-
-    override val name: String = "Tree"
-    override fun clone(): ExecValue {
-        return BinaryTreeValue(manimObject, value, style, animatedStyle)
-    }
-
-    override fun toString(): String {
-        return "$value"
-    }
-}
-
-data class Array2DValue(override var manimObject: MObject, val array: Array<Array<ExecValue>>, var style: StyleProperties = StyleProperties(), var animatedStyle: AnimationProperties? = null) : ExecValue() {
-    override val value: Array<Array<ExecValue>> = array
-    override val name: String = "Array"
-    override fun clone(): ExecValue {
-        return Array2DValue(manimObject, array, style, animatedStyle)
-    }
-
-    override fun toString(): String {
-        return array.joinToString(
-            ", ", "[", "]",
-            transform = {
-                it.joinToString(", ", "[", "]")
-            }
-        )
     }
 }
