+++ src/main/kotlin/com/manimdsl/runtime/utility/UtilityFunctions.kt
@@ -1,9 +1,10 @@
 package com.manimdsl.runtime.utility
 
 import com.manimdsl.frontend.*
-import com.manimdsl.linearrepresentation.DataStructureMObject
-import com.manimdsl.linearrepresentation.InitTreeStructure
-import com.manimdsl.runtime.*
+import com.manimdsl.runtime.BoolValue
+import com.manimdsl.runtime.CharValue
+import com.manimdsl.runtime.DoubleValue
+import com.manimdsl.runtime.ExecValue
 import com.manimdsl.stylesheet.PositionProperties
 
 private const val WRAP_LINE_LENGTH = 50
@@ -57,50 +58,11 @@ fun wrapString(text: String, max_length: Int = WRAP_LINE_LENGTH): String {
     return sb.toString()
 }
 
-private fun makeExpressionNode(value: ExecValue, lineNumber: Int): ExpressionNode {
+fun makeExpressionNode(value: ExecValue, lineNumber: Int): ExpressionNode {
     return when (value) {
         is CharValue -> CharNode(lineNumber, value.value)
         is DoubleValue -> NumberNode(lineNumber, value.value)
         is BoolValue -> BoolNode(lineNumber, value.value)
         else -> VoidNode(lineNumber)
-    }
-}
-
-fun makeConstructorNode(assignedValue: ExecValue, lineNumber: Int): ConstructorNode {
-    return when (assignedValue) {
-        is ArrayValue -> {
-            val dim = NumberNode(lineNumber, assignedValue.array.size.toDouble())
-            val initialiser = DataStructureInitialiserNode(
-                assignedValue.array.map { makeExpressionNode(it, lineNumber) }
-            )
-            val type = (assignedValue.manimObject as DataStructureMObject).type
-            ConstructorNode(lineNumber, type, listOf(dim), initialiser)
-        }
-        is Array2DValue -> {
-            val dimY = NumberNode(lineNumber, assignedValue.array.size.toDouble())
-            val dimX = NumberNode(lineNumber, assignedValue.array[0].size.toDouble())
-            val initialiser = Array2DInitialiserNode(
-                assignedValue.array.map {
-                    it.map { v -> makeExpressionNode(v, lineNumber) }
-                }
-            )
-            val type = (assignedValue.manimObject as DataStructureMObject).type
-            ConstructorNode(lineNumber, type, listOf(dimY, dimX), initialiser)
-        }
-        is StackValue -> {
-            val type = (assignedValue.manimObject as DataStructureMObject).type
-            val initialiser = DataStructureInitialiserNode(
-                assignedValue.stack.map { makeExpressionNode(it, lineNumber) }
-            )
-            ConstructorNode(lineNumber, type, listOf(), initialiser)
-        }
-        is BinaryTreeValue -> {
-            with(assignedValue.manimObject as InitTreeStructure) {
-                val internalType = this.type.internalType
-                val root = ConstructorNode(lineNumber, NodeType(internalType), listOf(makeExpressionNode(assignedValue.value.value, lineNumber)), EmptyInitialiserNode)
-                ConstructorNode(lineNumber, TreeType(internalType), listOf(root), EmptyInitialiserNode)
-            }
-        }
-        else -> ConstructorNode(lineNumber, ArrayType(NullType), listOf(), EmptyInitialiserNode)
     }
 }
